small tuning adjustments

+++ fit_2exp_Pmu_production.m
@@ -21,7 +21,7 @@ L1 = -1./pf1(1);
 % Starting guess
 options = optimset('MaxFunEvals',50000);
 x0 = [P1/1.5 P1/3 L1/2 L1.*1.5]; %[neg_P fast_P neg_L fast_L]
-xopt = fminsearch(@(x) sum(((x(1).*exp(-zmass./x(3)) + x(2).*exp(-zmass./x(4)))-Pmu).^2),x0,options);
+xopt = fminsearch(@(x) sum(((x(1).*exp(-zmass./x(3)) + x(2).*exp(-zmass./x(4)))-double(Pmu)).^2),double(x0),options);
 
 % assign output
 Pnmu = xopt(1);
